1. 블록체인 연동(127.0.0.1 을 상대 ip로 확인함) 2. 전체적인 코드 정리 3. 겹치는 기능들은 DataClass로 통일 시킴

+++ src/main/java/GUI/MyCalendar.java
@@ -74,10 +74,9 @@ class MyCalendar extends Panel {
       //JSONObject data = (JSONObject) json.get(key);
 
       System.out.println(data);
-      cyear.select((int) data.get("year")-2010);
-      cmonth.select((int)data.get("month")-1);
-      cdate.select((int)data.get("date")-1);
-
+      cyear.select((int) (long) data.get("year")-2010);
+      cmonth.select((int) (long) data.get("month")-1);
+      cdate.select((int) (long)data.get("date")-1);
    }
 
    MyCalendar() {
